LaBuena refactor Change data type for location to bigdecimal"

+++ LaBuenaServices/src/main/java/mx/com/labuena/services/tos/Location.java
@@ -1,34 +1,36 @@
 package mx.com.labuena.services.tos;
 
+import java.math.BigDecimal;
+
 /**
  * Created by moracl6 on 8/1/2016.
  */
 
 public class Location {
-    private double latitude;
-    private double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 
     public Location() {
     }
 
-    public Location(double latitude, double longitude) {
+    public Location(BigDecimal latitude, BigDecimal longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public double getLatitude() {
+    public BigDecimal getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(double latitude) {
+    public void setLatitude(BigDecimal latitude) {
         this.latitude = latitude;
     }
 
-    public double getLongitude() {
+    public BigDecimal getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(double longitude) {
+    public void setLongitude(BigDecimal longitude) {
         this.longitude = longitude;
     }
 }
